Refactor ToolLibraryPopoverView and related components for improved state management and modularity
- Removed unused properties and state variables from ToolLibraryPopoverView.
- Introduced ToolLibraryStorePublisher to handle store publishing logic and state.
- Created ToolLibraryStorePublishSheetView for better separation of concerns in the publishing UI.
- Updated methods to utilize the new ToolLibraryStorePublisher for managing published store apps and publishing processes.
- Enhanced error handling and state management during publishing operations.
- Added unit tests for new functionalities and updated existing tests to reflect changes in the store publishing logic.
