+++ Ironsmith/Features/ToolLibrary/Popover/ToolLibraryPopoverView.swift
@@ -2,7 +2,6 @@ import AuthenticationServices
 import Foundation
 import SwiftData
 import SwiftUI
-import UniformTypeIdentifiers
 
 struct ToolLibraryPopoverView: View {
     @Environment(\.modelContext) private var modelContext
@@ -20,22 +19,11 @@ struct ToolLibraryPopoverView: View {
     #endif
     let appUpdateStore: AppUpdateStore
     private let welcomeOnboardingStore: WelcomeOnboardingStore
-    private let storeClient: IronsmithStoreClient
-    private let iconClient: ToolIconClient
     @State private var toolLibraryStore = ToolLibraryStore()
+    @State private var storePublisher: ToolLibraryStorePublisher
     @State private var toolPendingDeletion: Tool?
     @State private var toolPendingRename: Tool?
     @State private var pendingRenameName = ""
-    @State private var publishedStoreAppsByID: [String: StoreAppListing] = [:]
-    @State private var publishingToolID: UUID?
-    @State private var publishName = ""
-    @State private var publishDescription = ""
-    @State private var publishDisplayName = ""
-    @State private var publishScreenshotData: Data?
-    @State private var publishScreenshotName: String?
-    @State private var isShowingPublishSheet = false
-    @State private var isPublishingToStore = false
-    @State private var storeErrorMessage: String?
     @State private var hasCheckedWelcomeOnboarding = false
     @State private var isShowingWelcomeOnboarding = false
     @State private var isShowingModelPicker = false
@@ -46,8 +34,7 @@ struct ToolLibraryPopoverView: View {
     init() {
         appUpdateStore = AppUpdateStore()
         welcomeOnboardingStore = WelcomeOnboardingStore()
-        storeClient = .live
-        iconClient = .live()
+        _storePublisher = State(initialValue: ToolLibraryStorePublisher())
     }
 
     @MainActor
@@ -59,11 +46,17 @@ struct ToolLibraryPopoverView: View {
     ) {
         self.appUpdateStore = appUpdateStore
         self.welcomeOnboardingStore = welcomeOnboardingStore ?? WelcomeOnboardingStore()
-        self.storeClient = storeClient ?? .live
-        self.iconClient = iconClient
+        _storePublisher = State(
+            initialValue: ToolLibraryStorePublisher(
+                storeClient: storeClient ?? .live,
+                iconClient: iconClient
+            )
+        )
     }
 
     var body: some View {
+        @Bindable var storePublisher = storePublisher
+
         // The menu bar popover stays intentionally small: tool list first, prompt last.
         VStack(spacing: 14) {
             ToolLibraryPopoverHeaderView(
@@ -175,7 +168,10 @@ struct ToolLibraryPopoverView: View {
                 await toolLibraryStore.refreshRestoreAvailability(for: tools)
             }
             .task(id: publishedStoreLinkRefreshID) {
-                await refreshPublishedStoreAppIDs()
+                await storePublisher.refreshPublishedStoreApps(
+                    isSignedIn: inferenceStore.ironsmithSession != nil,
+                    tools: tools
+                )
             }
 
             PromptComposerView(
@@ -278,7 +274,7 @@ struct ToolLibraryPopoverView: View {
         ) {
             Button("OK", role: .cancel) {}
         } message: {
-            Text(storeErrorMessage ?? "")
+            Text(storePublisher.errorMessage ?? "")
         }
         .confirmationDialog(
             "Delete App?",
@@ -319,7 +315,7 @@ struct ToolLibraryPopoverView: View {
                 onComplete: completeWelcomeOnboarding
             )
         }
-        .sheet(isPresented: $isShowingPublishSheet) {
+        .sheet(isPresented: $storePublisher.isShowingPublishSheet) {
             storePublishSheet
         }
         .sheet(isPresented: $isShowingModelPicker) {
@@ -329,27 +325,37 @@ struct ToolLibraryPopoverView: View {
 
     @ViewBuilder
     private var storePublishSheet: some View {
-        if let tool = tools.first(where: { $0.id == publishingToolID }) {
+        @Bindable var storePublisher = storePublisher
+
+        if let tool = tools.first(where: { $0.id == storePublisher.publishingToolID }) {
             ToolLibraryStorePublishSheetView(
                 tool: tool,
                 isUpdatingPublishedListing: canUpdateStoreVersion(for: tool),
-                publishName: $publishName,
-                publishDescription: $publishDescription,
-                publishDisplayName: $publishDisplayName,
-                publishScreenshotName: publishScreenshotName,
-                needsDisplayName: needsStoreDisplayName,
-                isPublishing: isPublishingToStore,
+                publishName: $storePublisher.publishName,
+                publishDescription: $storePublisher.publishDescription,
+                publishDisplayName: $storePublisher.publishDisplayName,
+                publishScreenshotName: storePublisher.publishScreenshotName,
+                needsDisplayName: storePublisher.needsDisplayName(inferenceStore: inferenceStore),
+                isPublishing: storePublisher.isPublishing,
                 onSaveDisplayName: {
-                    Task { await saveStoreDisplayName() }
+                    Task { await storePublisher.saveDisplayName(inferenceStore: inferenceStore) }
                 },
                 onChooseScreenshot: { url in
-                    importStoreScreenshot(from: url)
+                    storePublisher.importScreenshot(from: url)
                 },
                 onCancel: {
-                    isShowingPublishSheet = false
+                    storePublisher.isShowingPublishSheet = false
                 },
                 onPublish: {
-                    Task { await publishToolToStore(tool) }
+                    Task {
+                        await storePublisher.publish(
+                            tool,
+                            modelContext: modelContext,
+                            inferenceStore: inferenceStore,
+                            defaultSettings: defaultGenerationSettings,
+                            routeStore: routeStore
+                        )
+                    }
                 }
             )
         }
@@ -393,172 +399,7 @@ struct ToolLibraryPopoverView: View {
     }
 
     private func canUpdateStoreVersion(for tool: Tool) -> Bool {
-        guard let storeAppId = tool.storeAppId else { return false }
-        return publishedStoreAppsByID[storeAppId] != nil
-    }
-
-    private func refreshPublishedStoreAppIDs() async {
-        guard inferenceStore.ironsmithSession != nil else {
-            publishedStoreAppsByID = [:]
-            return
-        }
-        let storeIDs = Set(
-            tools.compactMap { tool -> String? in
-                guard tool.storeAppId != nil else { return nil }
-                return tool.storeId
-            }
-        )
-        guard !storeIDs.isEmpty else {
-            publishedStoreAppsByID = [:]
-            return
-        }
-
-        do {
-            var ownedAppsByID: [String: StoreAppListing] = [:]
-            for storeID in storeIDs {
-                var cursor: String?
-                repeat {
-                    let page = try await storeClient.listApps(storeID, .mine, nil, cursor)
-                    for app in page.apps {
-                        ownedAppsByID[app.id] = app
-                    }
-                    cursor = page.nextCursor
-                } while cursor != nil
-            }
-            publishedStoreAppsByID = ownedAppsByID
-        } catch {
-            publishedStoreAppsByID = [:]
-        }
-    }
-
-    private var needsStoreDisplayName: Bool {
-        (inferenceStore.ironsmithAccountSummary?.profile?.displayName ?? "")
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-            .isEmpty
-    }
-
-    private func linkedPublishedApp(for tool: Tool) -> StoreAppListing? {
-        guard let storeAppId = tool.storeAppId else { return nil }
-        return publishedStoreAppsByID[storeAppId]
-    }
-
-    private func beginPublishingToStore(_ tool: Tool) async {
-        await inferenceStore.refreshIronsmithAccountSummary()
-        guard inferenceStore.ironsmithSession != nil else {
-            storeErrorMessage = "Sign in with Ironsmith before publishing to the App Store."
-            return
-        }
-        await refreshPublishedStoreAppIDs()
-        publishingToolID = tool.id
-        publishName = tool.name
-        publishDescription = linkedPublishedApp(for: tool)?.description ?? "Created with Ironsmith."
-        publishDisplayName = inferenceStore.ironsmithAccountSummary?.profile?.displayName ?? ""
-        publishScreenshotData = nil
-        publishScreenshotName = nil
-        isShowingPublishSheet = true
-    }
-
-    private func saveStoreDisplayName() async {
-        let trimmed = publishDisplayName.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return }
-        do {
-            _ = try await storeClient.updateProfileDisplayName(trimmed)
-            await inferenceStore.refreshIronsmithAccountSummary()
-            publishDisplayName = inferenceStore.ironsmithAccountSummary?.profile?.displayName ?? trimmed
-        } catch {
-            storeErrorMessage = IronsmithErrorPresentation.message(for: error)
-                ?? error.localizedDescription
-        }
-    }
-
-    private func publishToolToStore(_ tool: Tool) async {
-        guard !isPublishingToStore else { return }
-        isPublishingToStore = true
-        defer { isPublishingToStore = false }
-
-        do {
-            if needsStoreDisplayName {
-                _ = try await storeClient.updateProfileDisplayName(
-                    publishDisplayName.trimmingCharacters(in: .whitespacesAndNewlines)
-                )
-                await inferenceStore.refreshIronsmithAccountSummary()
-            }
-
-            let source = try String(
-                contentsOf: try tool.packageLayout.packageFileURL(for: tool.contentViewSourcePath),
-                encoding: .utf8
-            )
-            let settings = tool.generationSettings(defaults: defaultGenerationSettings)
-            let app: StoreAppListing
-            if let linkedApp = linkedPublishedApp(for: tool) {
-                app = try await storeClient.publishVersion(
-                    StoreVersionPublicationRequest(
-                        storeId: linkedApp.storeId,
-                        appId: linkedApp.id,
-                        sourceCode: source,
-                        generationSettings: settings,
-                        iconPNG: nil,
-                        screenshotPNGs: publishScreenshotData.map { [$0] } ?? [],
-                        replaceScreenshots: publishScreenshotData != nil,
-                        remixedFromVersionId: tool.storeRemixedFromVersionId
-                    )
-                )
-            } else {
-                _ = try await iconClient.ensureIconAssets(
-                    ToolIconRequest(displayName: tool.name, layout: tool.packageLayout)
-                )
-                let iconPNG = try Data(contentsOf: tool.packageLayout.cachedAppIconPNGURL)
-                app = try await storeClient.publishApp(
-                    StorePublicationRequest(
-                        storeId: tool.storeId ?? IronsmithStoreConstants.communityStoreId,
-                        name: publishName.trimmingCharacters(in: .whitespacesAndNewlines),
-                        description: publishDescription.trimmingCharacters(in: .whitespacesAndNewlines),
-                        sourceCode: source,
-                        generationSettings: settings,
-                        iconPNG: iconPNG,
-                        screenshotPNGs: publishScreenshotData.map { [$0] } ?? [],
-                        remixedFromVersionId: tool.storeRemixedFromVersionId
-                    )
-                )
-            }
-
-            applyPublishedStoreLinkage(app, to: tool)
-            try modelContext.save()
-            publishedStoreAppsByID[app.id] = app
-            isShowingPublishSheet = false
-            routeStore.open(.store(.publishedApp(app.id)))
-        } catch {
-            modelContext.rollback()
-            storeErrorMessage = IronsmithErrorPresentation.message(for: error)
-                ?? error.localizedDescription
-        }
-    }
-
-    private func importStoreScreenshot(from url: URL) {
-        let didAccess = url.startAccessingSecurityScopedResource()
-        defer {
-            if didAccess {
-                url.stopAccessingSecurityScopedResource()
-            }
-        }
-        do {
-            publishScreenshotData = try Data(contentsOf: url)
-            publishScreenshotName = url.lastPathComponent
-        } catch {
-            storeErrorMessage = IronsmithErrorPresentation.message(for: error)
-                ?? error.localizedDescription
-        }
-    }
-
-    private func applyPublishedStoreLinkage(_ app: StoreAppListing, to tool: Tool) {
-        tool.storeId = app.storeId
-        tool.storeAppId = app.id
-        tool.storeVersionId = app.currentVersion.id
-        tool.storeVersionNumber = app.currentVersion.versionNumber
-        tool.storeSourceSha256 = app.currentVersion.sourceSha256
-        tool.storeImportedAt = Date()
-        tool.storeRemixedFromVersionId = app.currentVersion.remixedFromVersionId
-        tool.updatedAt = Date()
+        storePublisher.canUpdateStoreVersion(for: tool)
     }
 
     private var canSubmitPrompt: Bool {
@@ -742,7 +583,13 @@ struct ToolLibraryPopoverView: View {
             isPromptFocused = focusPrompt
         case .publishTool(let id):
             guard let tool = tools.first(where: { $0.id == id }) else { return }
-            Task { await beginPublishingToStore(tool) }
+            Task {
+                await storePublisher.beginPublishing(
+                    tool,
+                    inferenceStore: inferenceStore,
+                    tools: tools
+                )
+            }
         }
     }
 
@@ -874,10 +721,10 @@ struct ToolLibraryPopoverView: View {
 
     private var storeErrorPresentedBinding: Binding<Bool> {
         Binding(
-            get: { storeErrorMessage != nil },
+            get: { storePublisher.errorMessage != nil },
             set: { isPresented in
                 if !isPresented {
-                    storeErrorMessage = nil
+                    storePublisher.errorMessage = nil
                 }
             }
         )
@@ -894,83 +741,6 @@ struct ToolLibraryPopoverView: View {
         )
     }
 }
-
-private struct ToolLibraryStorePublishSheetView: View {
-    let tool: Tool
-    let isUpdatingPublishedListing: Bool
-    @Binding var publishName: String
-    @Binding var publishDescription: String
-    @Binding var publishDisplayName: String
-    let publishScreenshotName: String?
-    let needsDisplayName: Bool
-    let isPublishing: Bool
-    let onSaveDisplayName: () -> Void
-    let onChooseScreenshot: (URL) -> Void
-    let onCancel: () -> Void
-    let onPublish: () -> Void
-    @State private var isChoosingScreenshot = false
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text(isUpdatingPublishedListing ? "Update Store Version" : "Publish to App Store")
-                .font(.headline)
-
-            if needsDisplayName {
-                VStack(alignment: .leading, spacing: 8) {
-                    TextField("Display Name", text: $publishDisplayName)
-                    Button("Save Display Name", action: onSaveDisplayName)
-                        .disabled(publishDisplayName.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty)
-                }
-            }
-
-            TextField("Name", text: $publishName)
-            TextField("Description", text: $publishDescription, axis: .vertical)
-                .lineLimit(3...5)
-
-            HStack {
-                Button {
-                    isChoosingScreenshot = true
-                } label: {
-                    Label("Screenshot", systemImage: "photo")
-                }
-                Text(publishScreenshotName ?? "No screenshot selected")
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-
-            HStack {
-                Spacer()
-                Button("Cancel", action: onCancel)
-                Button(isUpdatingPublishedListing ? "Update" : "Publish", action: onPublish)
-                    .buttonStyle(.borderedProminent)
-                    .disabled(!canPublish || isPublishing)
-                if isPublishing {
-                    ProgressView()
-                        .controlSize(.small)
-                }
-            }
-        }
-        .padding(18)
-        .frame(width: 340)
-        .fileImporter(
-            isPresented: $isChoosingScreenshot,
-            allowedContentTypes: [.image],
-            allowsMultipleSelection: false
-        ) { result in
-            if case .success(let urls) = result, let url = urls.first {
-                onChooseScreenshot(url)
-            }
-        }
-    }
-
-    private var canPublish: Bool {
-        !publishName.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-            && !publishDescription.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-            && (!needsDisplayName || !publishDisplayName.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty)
-            && !tool.name.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-}
-
 #Preview("Tool Library") {
     let container = try! IronsmithModelContainerFactory.make(isRunningTests: true)
     let menuBarPopoverPresentationStore = MenuBarPopoverPresentationStore()
