+++ IronsmithTests/Features/ToolLibrary/ToolLibraryTestSupport.swift
@@ -63,6 +63,13 @@ extension ToolLibraryTests {
                     recentLedger: []
                 )
             },
+            updateProfile: { _ in
+                IronsmithAccountProfile(
+                    id: "00000000-0000-4000-8000-000000000001",
+                    email: "jade@example.com",
+                    displayName: nil
+                )
+            },
             fetchCreditPacks: { [] },
             createCheckoutSession: { _ in
                 throw IronsmithAccountClientError.notConfigured
@@ -92,6 +99,13 @@ extension ToolLibraryTests {
             signOut: {},
             fetchAccountSummary: {
                 await fetchCapture.fetch()
+            },
+            updateProfile: { _ in
+                IronsmithAccountProfile(
+                    id: "00000000-0000-4000-8000-000000000001",
+                    email: "jade@example.com",
+                    displayName: nil
+                )
             },
             fetchCreditPacks: { [] },
             createCheckoutSession: { _ in
